C99-44/CTS-R3: union model completed and dispositioned — float puns shipped via new emitrust.bitcast, pun access surface hardened, both entries closed
Extends the one-slot union model with same-width float/int pun arms
(float/32-bit, double/64-bit): a new emitrust.bitcast dialect op renders
Rust's to_bits/from_bits (bit-exact by definition, deliberately distinct
from the value-converting `as` of emitrust.cast; verifier requires one
f32/f64 side and one same-width integer side). Access sites reinterpret
through a shared reinterpretScalarBits primitive; constant initializers
through a float-pun arm cross the domain at compile time (APFloat
bitcast) for globals.

Hardens the pun access surface for BOTH pun families:
- designated local initializers through a pun arm used to store the
  arm-typed value into the slot-typed field — uncompilable Rust
  (pre-existing defect, now fixed and pinned);
- compound assignment through a pun arm now reinterprets slot->arm,
  computes at the arm's type, and reinterprets back (a float arm would
  otherwise have VALUE-converted through convertScalarValue — silent
  miscompile);
- ++/-- reinterprets first, so a float pun arm lands on the general
  non-integer ++/-- rejection instead of raw bit arithmetic;
- value-position assignment/compound-assignment through a pun arm now
  reinterprets the re-loaded slot value (previously a located rejection).

Rejection matrix retained and fully pinned: bit-field/pointer/unnamed
arms, empty unions, differing-size scalars (int/int, float/long,
double/float), aggregate arms (new pin), byte-array arm access, float
pun-arm ++. test/Import/C/union.c (int/float rejection) deleted as
stale — that shape now imports; unions-invalid.c supersedes it, so the
lit total moves 265 -> 264.

design.md: C99-44 and CTS-R3 rewritten to one consistent shipped state
and both ticked; the CTS-R3 note calling 00218.c out of scope was stale
(it passes via single-arm union + C99-45 + CTS-P2); CTS-R2's
named-union sentence updated; range-refinement op inventory notes
bitcast as opaque/TOP.

check-emitrust: 264/264. Ledger unchanged: total=220 passed=217
miscompiled=0 unsupported=3 (00042, 00210, 00218 all PASS).

+++ test/Import/C/structs-anon-member-invalid.c
@@ -3,8 +3,8 @@
 // whose arms differ in type — or whose arm is wider than one flattened
 // field — stays rejected with the same located "union type" diagnostic
 // union types get everywhere (CTS-R3 territory); a named union member
-// whose arms fall outside the one-slot model (e.g. a float arm) keeps
-// that rejection too. A spelling collision between the parent and
+// whose arms fall outside the one-slot model (a float arm of a width no
+// integer arm shares) keeps that rejection too. A spelling collision between the parent and
 // an anonymous member's field is a clang error before import (C11
 // 6.7.2.1p13 puts both in one member namespace), also with a location.
 // RUN: split-file %s %t
@@ -15,7 +15,7 @@
 
 // MIXED: mixed-arms.c:{{[0-9]+}}:{{[0-9]+}}: error: unsupported: union type
 // WIDE: wide-arm.c:{{[0-9]+}}:{{[0-9]+}}: error: unsupported: union type
-// NAMED: named-union.c:{{[0-9]+}}:{{[0-9]+}}: error: unsupported: union type
+// NAMED: named-union.c:{{[0-9]+}}:{{[0-9]+}}: error: unsupported: union
 // COLLIDE: collision.c:{{[0-9]+}}:{{[0-9]+}}: error: member of anonymous struct redeclares 'x'
 
 //--- mixed-arms.c
@@ -46,10 +46,11 @@ struct s {
 struct s g;
 
 //--- named-union.c
-// A named union member with a non-integer arm stays outside the
-// one-slot aliasing model even after CTS-R3 named-union support.
+// A named union member whose arms fall outside the one-slot model (a
+// float arm against a WIDER integer arm; same-width pairs are admitted
+// as puns) stays rejected even after CTS-R3 named-union support.
 union u {
-  int a;
+  long a;
   float b;
 };
